Start with the program entrypoint
Define the experience I'd like to support. Each HTTP handler should be
configurable. Defaults should serve the 80-90% case, but when user
wishes to provide their own `POST /sign-in` handler, it can be set in
main.go

At the time of commiting, this interface does not exist, but it's an
ideal.